requirements.txt
Django==2.2
django-extensions>=2
djangorestframework==3.7.7
django-crispy-forms==1.7.0

django-widget-tweaks
mcstatus
django-ckeditor
django-jet
Pillow
django-admin-honeypot
gunicorn
discord.py==0.13.0
praw==3.3.0
psutil==3.3.0
pytz==2016.7
tabulate==0.7.5
django-import-export==1.2.0
python-dateutil==2.5.3
parsedatetime==2.1
bpython
psycopg2-binary
